BUG#36400360 removeCluster() fails with error on group_replication_enable_member_action
When the Primary instance of a Replica Cluster is changed, with for
example .setPrimaryInstance(), attempting to remove that Cluster from
the ClusterSet with .removeCluster() fails with the following error:

ERROR: Error enabling automatic super_read_only management at
<secondary>:<port>: MySQL Error 3910 (HY000): The function
'group_replication_enable_member_action' failed. Member must be the
primary or OFFLINE.

The root cause is that the call to the UDF
'SELECT group_replication_enable_member_action()' is being executed
on all reachable Cluster members. The UDF must be called on the Primary
member only. Group Replication will then propagate the change to all
Cluster members.

If the first member of that list of reachable Cluster members is a
secondary, the UDF will fail and cause the whole command to fail.

This patch fixes that by ensuring the UDF is called only on the Primary
member when the Cluster is ONLINE. If the Cluster, or target instance
is OFFLINE then the UDF can be called too. This is needed to revert the
creation of a Replica Cluster or to remove an OFFLINE Cluster from a
ClusterSet.

Change-Id: I044704c49fdaba4c2ef13c068a773f99bcceb509