Fix random fails in some AdminAPI tests
Also, when changing the primary Cluster of a ClusterSet, the AdminAPI
was updating the replica settings of the new primary multiple times
(once per secondary of the current primary cluster), which is wasteful
because it only needs to be made once. Change it so that it is only
called once.

Change-Id: I0732e97c2357abd6fa87bcaf7c12edad8534fa89