BUG#36394895 Add support to list objects for checks performed by MySQL shell upgrade checker
Issues for the upgrade checker include the affected database object
in the dbObject entry in the generated JSON document for each issue,
however, the object type is not properly defined.

This patch adds a new dbObjectType entry in the JSON output to
satisfy the requirement.

Additionally, this patch deletes the zerofill check which was
initially added by mistake (BUG#27927641) and which was just
unregistered but not removed.

Change-Id: Ib908b993bb85590aaa518664d5dcc94b57d8fc14