Bug#36422408,Bug#36422502,Bug#36422492 Command line password prompting fixes
- Bug#36422408 	--no-password doesn't work if password set in my.cnf

Specifying --no-password with other password options was conflicting
and erroring out. This is incompatible with password being specified in
~/.my.cnf, so must be made consistent with handling of other options
(in case of conflict, last one wins).

- Bug#36422492    Can't connect via mysqlsh with 'MySQL Error 1045
(28000): Access denied for user xxx@xxx (using password: YES)

- Bug#36422502 	-p option is not forcing password to be prompted

The -p and --password options were not being handled for the case where
no value is given, which indicates the password should be prompted.
Password was not prompted if password specified earlier in the cmdline
or in my.cnf, even if -p was specified.

Additionally, MFA password options (--password2 and --password3) were
changed to behave like --password and prompt for a password is no
value is given, but still not prompt for anything by default.

Refactored MFA passwords to not store an additional copy of password1,
as --password1 is just an alias for --password/-p.

Change-Id: I535836fe1e86b57128e1bc4dde8136d886275044