BUG#36405653 Empty object list in authUsage and pluginUsage checks
The Feature_life_cycle checks were grouping the affected items per
feaure/check in the issue description, which caused the affected
items to not be available as dbObjects in the JSON output as any of
the other checks.

This patch fixes the problem by removing the grouping from the check
itself and adding it into the formatter. Now, the Feature_life_cycle
checks will produce an issue per affected item an the grouping logic
will be a display matter only implemented in the Text formatter,
the JSON formatter is not affected but with the new logic the JSON
output will now properly render the affected dbObject.

Change-Id: Ib016b044c5e035106993ce206bff6553dc617118