WL#15820 BUG#36610582 BUG#36611243 BUG#36610397 AdminAPI: .execute
This patch adds a new command to Cluster, ClusterSet and ReplicaSet
objects: execute.

It allows the user to run a single query simultaneously in multiple
instances of the topology. The instances can be specified individually
or by using predetermined keywords such as "all" / "a", to target all
instances of the topology, or "read-replicas" / "rr" to target only
read-replicas, for example.

Change-Id: Id5894c4feec9c8e33ea36970ecffa6ffacb15099