BUG#35181181 rebootCluster* fails when GR stopped and repl channel reset on replica cluster
When rebooting a cluster or rejoining an instance, it's assumed that the
recovery user already exists, meaning that it isn't (re)created.
However, there are scenarios where this can happen, for example, if the
user executes a "RESET REPLICA ALL" on an instance.

This patch adds support for this scenario by assuming that the recovery
user does not exist, (re)creating it regardless.

Also included in this patch is an overall reorganization of the logic
responsible for create, managing and deleting replication accounts.

Change-Id: I255648b8ab151bc8bafe8561980ee64b6257ac76