BUG#36477603 Loading a table which contains a virtual column generates a warning
When loading a dump which contained a table with a virtual column into
an instance which supports BULK LOAD, a warning was reported for each
row in that table.

Table metadata written by the dumper contains a list of columns, for
which data was dumped to the data files; generated columns are not on
this list. When loading the data, loader uses this list to construct the
LOAD DATA statement. When table is checked for compatibility with BULK
LOAD, this list was cleared in a local variable. This local variable
used a reference to the global variable, which was also cleared. This
caused the loader to create a statement which didn't explicitly specify
the columns, hence server reported:
  error 1261: Row 1 doesn't contain data for all columns

The data structure has been reorganized, global variable is no longer
modified when a local variable is changed.

Change-Id: I015abd7bfb47ce9db1fe75d2463933b82d6eac8e