BUG#36380502 deprecate --dba=enableXProtocol
"--dba=enableXProtocol" was added as a helper command to install the
X (Protocol) Plugin on the target MySQL Server.
It was added to help users, when the X Plugin was not installed by
default, in MySQL 5.7.

With MySQL 8.0 the plugin comes installed by default.

MySQL 5.7 reached EOL in October 2023 so this patch deprecates the
functionality.

Also fixed a segmentation fault when executing the command without a
connection to the target instance.

Change-Id: If1b82a0bc601bb613eca06885c6d5e44d60a8eff