Bug#36502096	crash when passing python objects as an argument to a python plugin function
Passing a native Python object which does not get converted to a native shell
type (like set, Thread etc) to a Python plugin function would result in an
reference counting imbalance, eventually ending in a crash.

The cause was that py::convert() was assuming it owned the reference of
the converted PyObject, when it's actually just borrowed. Incrementing
the refcount fixes it.

Also, implemented a shortcut for the case of Python code calling Python plugins,
in order to avoid a double-conversion of the plugin arglist from Python to Value
and back to Python.

Change-Id: I77a470182f0bf2e0b4e9d60e779392748c3dc6b5