BUG#114707 Handle None diff values when collecting diagnostics
We have seen that the debug utility that collects diagnostics via MySQL shell
fails with an `AttributeError` when executed in a member of a broken GR
cluster. It seems that the problem is more generic, since this error happens
when the `dump_diff()` function is invoked with a `diff` object that contains
at least one value that equals None.

Fix this by adding a check to ensure that the value of every diff value to be
left-justified is not None before appending it to the current line.

Also, print a descriptive message to inform the user that a None diff value was
detected and will be skipped from the current line.

Signed-off-by: Ioannis Androulidakis <androulidakis.ioannis@gmail.com>