Enabling mysql_native_password by default on test sandboxes.
Not having the mysql_native_password results in random error messages
on the server when trying to connect using non existing accounts, this
results in random failures on the shell tests. This patch enables
the mysql_native_password plugin by default on sandboxes deployed using
the testutil.deploySandbox and testutil.deployRawSandbox functions.

Change-Id: I18e42b1e3bea985a87e1bf49179743bbcc8da954