BUG#36562916 dba.reboot_cluster_from_() disables SRO on primary member of INVALIDATED Cluster
An outcome of restoring an INVALIDATED Cluster using
dba.reboot_cluster_from_complete_outage() is that super_read_only is
disabled from its primary member.

Shell detects that a Cluster is INVALIDATED, because it can either
reach the ClusterSet's metadata on the other Clusters or its own
metadata indicates that, and does not automatically rejoin the instances
back to the Cluster to avoid introducing errant transactions, however,
it does not disable GR's member action
mysql_disable_super_read_only_if_primary to ensure super_read_only is
NOT disabled on the Cluster's primary allowing clients to perform
updates and introduce errant transactions.

This patch fixes that by ensuring the member action is disabled on such
scenario.

Change-Id: I3418f8d22fd2124c3e7e837c3608973f74ccfb02