WL#15378 BUG#34564242 BUG#36442258 BUG#36652856 BUG#36654761 BUG#36663491 AdminAPI - ClusterSet dissolve
This patch adds a new command to ClusterSet: dissolve. It allows the
user to completely dissolve a ClusterSet, keeping the user data intact.
This essentiality dissolves the ClusterSet and all Clusters that are
part of it.

This patch also adds a new parameter, "dissolve", to
clusterset.removeCluster, that allows the user to choose if the cluster
being removed is dissolved or not. Before this patch, the cluster was
always dissolved, but now, with this new option, the user can choose to
keep the cluster as a standalone.

Finally, cluster.dissolve and replicaset.dissolve were slightly modified
so that any error found while syncing instances and removing then (after
metadata has already been removed from the primary), no longer causes
the command to stop, and continues on to the next instance. This is
because the action of removing the metadata is irreversible: if
something fails after that (i.e.: instances syncing and leaving the
cluster), nothing can be done to reverse it. As such, stopping if an
instance throws an error at this stage only creates more inconsistent
instances. Now, by allowing the command to proceed, all remaining
instances are processed, regardless of any error that might have
occurred with any previous instance.

Change-Id: I4dff2d79daab5ceeddd5ddb1778834d9f09a682a