BUG#36493316 Shell dump fails on Primary Key check (ocimds:true) for PK with ENUM column
When running a dump with the ocimds:true option, tables with primary
keys defined on an ENUM column where reported as missing the PK, and
dump failed.

For each table, dumper stored index which was selected to be used for
chunking, alongside an information on whether this index is a primary
key. As a fix for BUG#35180061, indexes on ENUM columns are no longer
selected for chunking. This meant that if a PK had an ENUM column, it
was not used for chunking and, as a side effect, dumper reported such
table as one which is missing a primary key.

As a fix, information about the index selected for chunking and whether
table has a primary key has been separated.

Change-Id: I52f44a2e5c7e591b7c5eb75d629e281f22ae06ce