BUG#36340714 Contribution: Fix handling of table/schema names containing non-ASCII chara ...
This commit corrects unicode detecting in schema_inconsistency check in upgrade checker and adds simple test for it.

Original contribution:
"It appears that the intention of this change was to ignore table names
containing non-ASCII characters, since their names are represented
differently in `information_schema.tables` and in
`information_schema.innodb_sys_tables`.  In the latter, disallowed
characters are replaced with '@' plus 4 hexadecimal digits representing the
character's Unicode codepoint.

However, the regular expression used for this (`@[0-9]`) was incorrect and
incomplete: it will *only* exclude Unicode codepoints 0x0000-0x9ffff, and
*won't* exclude Unicode codepoints 0xa000-0xffff. This commit fixes the
mistake in the regular expression.

In the longer term, tables whose names contain non-ASCII characters should
*not* be excluded from schema inconsistency checks.

This contribution is under the OCA signed by Amazon and covering
submissions to the MySQL project."

Original contributor: Daniel Lenski (Github user: dlenski)

Change-Id: I42f33248f849610b3a92ac321229365080e0a8ac