BUG#36470302 mysqlsh util.dumpSchemas does not work when HYPERGRAPH is enabled
Dumper executes EXPLAIN statements when chunking the tables. The output
format was not explicitly specified, and dumper expected TRADITIONAL
(tabular) format.

When the HYPERGRAPH is enabled, the TRADITIONAL format cannot be used,
and the default format of EXPLAIN statement changes to TREE, which
caused the error:
  Util.dumpSchemas: invalid field name rows (ArgumentError)

As a fix, the output format of executed EXPLAIN statements is explicitly
set to JSON.

Change-Id: I4d9aaaacddbd66d8258646972ab664572a1aade1