BUG#36433418 MySQL Shell doesn't prompt for password for '-p' - postfix
Using testutil.call_mysqlsh will use the binaries on the standard shell
packages when the tests are executed in direct mode (or when a test is
sufixed with _norecord). Since the plaintext credential helper is not
part of the standard package, this caused a hung and so a bad termination
of the tests in all platforms.

The solution is to force using binaries from the test package by using
the "mysqlshrec" as the binary to execute in the call.

Change-Id: I6d2e6b8dd119013724abf25d5576b4f347ae32de