BUG#36589677 collect_diagnostics() fails with AttributeError when handling None diff values
util.collect_diagnostics() fails with an `AttributeError` when running
against an InnoDB Cluster on which the Group Replication plugin was
uninstalled from one or multiple members.

The error happens when the dump_diff() function is invoked with a diff
object that contains at least one value that is 'None' (unexisting). The
root cause in this case is that the utility attempts to gather the
values for Group Replication sysvars which will not exist when the
plugin is uninstalled.

This patch fixes that by adding a check to ensure that the value of each
single diff value is not None before appending it to the line being
proecessed.

Based on a code contribution from Ioannis Androulidakis, tracked by:
Bug#114707.

Change-Id: Ice2b23f8e0c4ba197957a9b54b89fa4eaecaa348