BUG#36495756 Admin API automatic fix metadata inconsistencies
Problem
-------
When the action of creating a new Replica Cluster
(with .create_replica_cluster()) fails due to a timeout and the revert
operation fails also because of a timeout, there's a chance that the
Replica Cluster becomes dangling, i.e. it's ONLINE but it's not
associated with the ClusterSet's metatada.
It becomes hard for the end user to understand what happened and any
attempt to retry the operation fails with confusing errors stating that
the target instance already belongs to a Cluster. Attempting to remove
the Cluster from the ClusterSet also fails with errors indicating that
the Cluster does not belong to the ClusterSet.

In this particular bug, the causes for the failure were low values of
'wait_timeout' and 'interactive_timeout'.

Analysis
--------
All metadata updates are atomic, to ensure consistency. On top of that,
the AdminAPI attempts to revert any failing operation to ensure the
instances and metadata are reverted to the original state. To achieve
that, all Metadata updates are executed inside transactions. For
example:

  - Create topology (Cluster, ReplicaSet, ClusterSet)
  - Add instances to the topology
  - Remove instances from the topology
  - Etc.

However, there is one exception:

  - Create a Replica Cluster in a ClusterSet

That particular operation is split into 2 actions+transactions:

  1. Create a Cluster and its metadata
  2. Add that Cluster to the ClusterSet's metadata (tables, views, etc.)

If one of those actions fails and the changes are not reverted, the
Metadata becomes inconsistent.

Solution
--------
1. Session timeouts

The AdminAPI creates and manages sessions to MySQL instances using a
connection pool. Every connection established by the AdminAPI will be
"prepared", i.e. a set of settings are configured to ensure the optimal
and correct behavior. For example, to ensure that concurrent command
calls do not result in errors due to higher consistency levels, all
sessions set the consistency level 'EVENTUAL'.

Considering users may have set lower values of timeout as seen in this
bug, this patch extends prepare_session() to set the value of
'wait_timeout' to the default value of 8 hours, if the currently used
value is lower than that.
This ensures no command and/or revert operation fails due to a timeout.

2. .rescan()

Because users may have Clusters broken by this bug when using older
Shell versions and also because there might be other causes of the
failures reported in the analysis, Cluster.rescan() was extended to
detect and offer the possibility of resolving Metadata inconsistencies.

The command now checks for Metadata inconsistencies and prompts the
user to take action and resolve those, when detected. A new option was
added to enable that action when in non-interactive mode or to override
the prompt. The option is called 'repairMetadata' and is disabled by
default.

3. Clean ClusterSet replication channel in Cluster.dissolve()

It's not possible to dissolve a Cluster that belongs to a
ClusterSet, the Cluster needs to be removed from the ClusterSet
(which will ensure the replication channel is reset among other
reconfiguration). However, due to the scenario described in this bug,
users should be able to dissolve a dangling Cluster and that should
ensure that the ClusterSet replication channel is removed too.
Otherwise, when trying to re-use the ex-members of that Cluster to
create a new Replica Cluster (or other) the command will fail with an
error indicating that no replication channels can be configured in the
instance.

For that reason, we should aim to clean as much as possible and
considering that the remove_channel() function is a no-op in case the
channel doesn't exist, we always call it starting with this patch.

Change-Id: Ibb456702f8d52fff9526f82bf3f929b0d94dca56