BUG#36014067 log filtering
When the sqlLog option was set to error, the failed SQL statement
was being logged without filtering considering the unsafe statements.
This patch fixes this by filtering out statements that match the
patterns defined in the logSql.ignorePatternUnsafe shell option.

Additionally, the filtered statements logging has been updated to
include the pattern that caused the filter to happen:

SQL: <filtered:<pattern>>
Change-Id: I95e4398ed1b5b6fefd8fe5da7beb51bf4974b1d5