BUG#36433418 MySQL Shell doesn't prompt for password for '-p' - postfix 2
When a connection is being attempted reusing the global session,
the credential manager is being used to read the passwords even they
are already present in the connection options. This patch reverts the
previous logic and noe deciding about the usage of the credential manager
is done only for the initial connection and based on presense of the -p
option.

Change-Id: Ifc6e75295d56435bd39c92dc29c2d47f9cfc6020