Removing code to delete mysqlsh from the modules dictionary.
The first time the mysqlsh module is imported in both C++ and Python
the loaded module points to the same memory. When the shell is reset
(i.e. in the unit tests) the same scenario will lead to 2 different
objects in memory:

- The module imported in C++ pointing to a new memory location
- The module imported in Python pointing to the original location

When the Python_context is terminated, it sets all the globals to None.
When the Python_context is re-created it sets the globals again.

Due to the inconsistency above, even the mysqlsh.globals.shell object
is set back the second time in C++, it is not done in the original
mysqlsh object to be use din Python, which causes it to be None in
the plugin loading context, which leads to errors loading plugins.

Change-Id: I7fce743dec0d70a6dd4eb73ee1139c5fba31138e