BUG#36658194 Running util.debug.collectDiagnostics w/ schemaStats:true get MySQL Error 1242
The diagnostic query that gathers information regarding the top biggest
tables on the server, doesn't take into account tables that might have
more than one column of type BLOB.

This patch fixes the query so that it prints an array of columns (if any
exist) to support multiple BLOB columns.

Also, the query now prints the engine type per-table, as well as
supporting TINYBLOB, MEDIUMBLOB and LONGBLOB column types (instead
of just BLOB).

Change-Id: I857308661606fe3bc6ec2923944dc0b110044532