WL#16408 BUG#36425775 BUG#36682841 BUG#36682800 BUG#36679360 BUG#36678689 BUG#36678522 BUG#36679512 BUG#36684566 BUG#36679921
AdminAPI: Replication compatibility checks

Problem
-------
The AdminAPI lacked validation for version compatibility during
asynchronous replication setup (in ReplicaSet, Read-Replicas, and
ClusterSet).

Context
-------
In general, replication is only fully supported from one release series
to the next higher release series. For example, 8.0.36 to 8.0.37, or
8.2.0 to 8.3.0.

With the LTS release model, the rule is that it is possible to
replicate from an LTS or Innovation release to:

  - The next LTS release
  - Any future Innovation release up until the next LTS release:
    LTS 8.4 -> LTS 9.7, but not LTS 8.4 -> LTS 10.8

1. Supported upgrade paths:

  - Within an LTS or Bugfix series
  - From an LTS or Bugfix series to the next LTS series
  - From an LTS or Bugfix release to an Innovation release before the
    next LTS series
  - From an Innovation series to the next LTS series
  - From within an Innovation series

2. Downgrade support:

For downgrade purposes, it's only fully supported to replicate from a
higher release series to a lower release series both are within the
same LTS release and only patch versions differ. For example:
8.4.20 -> 8.4.11.

Finally, it's possible, although limited to rollback operations, to
replicate:

  - From an LTS or Bugfix series to the previous LTS or Bugfix series
  - From an LTS or Bugfix series to an Innovation series after the
    previous LTS series
  - From within an Innovation series

Patch details
-------------
This patch adds replication compatibility checks in the AdminAPI,
informing users of the following scenarios:

  - Versions are compatible
  - Versions are incompatible: command blocked
  - Versions have limited compatibility to downgrade/rollback operations

The checks are performed during asynchronous replication channel setup
and in the .status() commands for Cluster (Read-Replicas), ReplicaSet,
and ClusterSet. These checks ensure compatibility for the direct source
and replica, as well as all potential donors/replicas when using a
channel with automatic failover (Read-Replicas, ClusterSet).

A new boolean shell option named 'dba.versionCompatibilityChecks' was
also added to control whether the checks result in errors and print
warnings, or are ignored with just log a message. This option is enabled
by default.

Additionally, it fixes a bug in .add_replica_cluster() and
.rejoin_instance() (for a read-replica) and the 'recoveryMethod' is set
to 'incremental' and mixed versions are used. Up to now, the commands
would perform a check to ensure clone compatibility in that scenario too
when they should not, resulting in errors when mixed versions are used
(because Clone does not support it).

Change-Id: Ie9b00632b1a024c1ce8847c30056b8a9e0f82f3c