WL#16003 BUG#34891382 BUG#36105235 Add support for multipart uploads using PARs
The backend code which uploads files to the OCI using prefix PARs has
been updated to use multipart uploads.

Previously, each created file was cached locally, and uploaded only
after the whole data was written. With this change, files are uploaded
while they are being created, in parts of 128MB.

BUG#34891382 Dump fails if an empty table is dumped when compression is set to none

When dumping the data to an OCI bucket, if there was an empty table and
compression was disabled, dump would fail.

The upload code was not uploading empty files, this has been fixed.

BUG#36105235 Uri_parser fails to parse an URI which contains unescaped '@' character

When an URL contained an unescaped '@' character in the target path,
i.e.:
  https://mysql.com/@.manifest.json
Uri_parser treated "mysql.com/" as a user name. Slash is not a valid
character in a user name, and an exception was thrown.

Change-Id: I60e7b656fc7b8fbc7b48131238ed73955e7edcdb