BUG#36658194 Running util.debug.collectDiagnostics w/ schemaStats:true get MySQL Error 1242 [postfix]
In some platforms, "sql_mode=only_full_group_by" is the default and the
query updated in the previous patch for this BUG fix fails because the
column 'blobs.column's is not aggregated and cannot be used in the
GROUP BY clause.

This postfix corrects the query to ensure aggregation is done for all
colums.

Change-Id: Id37982cfccc15a91dd1a8d34fbb5a23deabea2c6