BUG#36701854 util.DumpInstance throws syntax error when using MySQL Shell 8.0.37 and the MySQL server is in 8.4 or a superior version
When using Shell 8.0.37 to dump instance running 8.4.0 server, operation
fails with:
  You have an error in your SQL syntax; check the manual that
  corresponds to your MySQL server version for the right syntax to use
  near 'MASTER STATUS' at line 1 (MYSQLSH 1064)

The 8.2.0 server has removed the 'SHOW MASTER STATUS' statement and
replaced it with the 'SHOW BINARY LOG STATUS'. This change is not
handled by the 8.0.X versions of Shell.

Given that newer versions of server may introduce breaking changes,
a dump from an instance which is running a server with a greater minor
version than the Shell version now fails with an error which suggests to
upgrade Shell.

+--------+--------+--------+
| Shell  | Server | Result |
+--------+--------+--------+
| 8.0.39 | 8.0.40 | OK     |
| 8.0.39 | 8.1.0  | FAIL   |
| 8.1.0  | 8.2.0  | FAIL   |
| 8.4.2  | 8.4.3  | OK     |
| 8.4.2  | 9.0.0  | FAIL   |
+--------+--------+--------+

Change-Id: I46bd9790962a2d0d3ce8d0fdf30443139e919a5c