BUG#36858205 Shell crashes when pressing a tab to complete a \command
Shell crashed when pressing a tab to complete a '\command ' (any
backslash command followed by a space).

The code which splits the entered line into the command itself and its
arguments (everything that's past the first sequence of whitespace
characters) was not handling the case where line ended with a
whitespace, this has been fixed.

Change-Id: I51d2f3280f81b4e64d7b7b8cd661d1e6e25fa5d4