Do not delete the OCI test bucket
We were experiencing random 404 errors when running the OCI tests, as a
workaround, the bucket is no longer deleted from within tests.

IOS-131241 / IOS-132306

This is happening because tests are creating and deleting the bucket with
the same name and so the web-server host might be having the stale
metadata for objectDB shard and thus its routed to incorrect objectDB shard.

Solutions:

    Either do not create / delete the same bucket name.
    Create/delete buckets with different names.

Change-Id: I545439247d799c716ec9526b4b71c666e0543541