BUG#36050341 BUG#36561962 Add `dropExistingObjects` option
BUG#36561962 Shell DUMP LOAD: No options to drop table/schema before restore

When loading a dump, if a user account or DDL object to be loaded
already exists in the target instance, loader reports an error. User
may set the `ignoreExistingObjects` loader option to true to ignore
these objects and load the dump anyway.

This patch introduces a new `dropExistingObjects` loader option, which
allows to drop existing user accounts and DDL objects (excluding
schemas) found in the dump. When resuming an interrupted load process,
objects which were completed in the previous run are not dropped.

The `dropExistingObjects` and `ignoreExistingObjects` options cannot be
both set to true.

The `dropExistingObjects` option cannot be set to true when the
`loadDdl` option is set to false.

BUG#36050341 copyInstance should drop created objects on error

A failed copy operation required manually dropping all the copied
objects, before retrying the operation.

This patch introduces a new `dropExistingObjects` copy option, which
allows to drop user accounts and DDL objects (excluding schemas) that
exist in the target instance, before copying them from the source
instance.

The `dropExistingObjects` and `ignoreExistingObjects` options cannot be
both set to true.

The `dropExistingObjects` and `dataOnly` options cannot be both set to
true.

Change-Id: I615d1b03df208dee96c6dd6422583476db0c4d79