BUG#36861912 MySQL Shell behavior on Console Exit
When running Shell via SSH (i.e. putty), if a command was entered in the
prompt and then the SSH connection was closed, it was possible that this
command was executed, even though the enter key was not pressed.

The linenoise module (which is used by Shell to read from terminal) was
not treating a situation where stdin is closed as an error. Instead,
it returned buffer holding the current contents of prompt, which was
then executed by Shell.

The linenoise module was adjusted to signal when stdin is closed, in
such case Shell aborts execution.

Change-Id: I494b0db0929f5b0b95d7fe871b95c4f4dc63246b