BUG#36652642 ClusterSet.createReplicaSet on 5.7 target should give speaking error message
AdminAPI commands targetting instances running MySQL 5.7 (e.g.
add_instance()) fail due to a query to an unexisting table
"performance_schema.user_defined_functions", resulting in an unhandled
error (1146).

The root cause was the Locking Service checker, which did not account
for absense of this table in MySQL 5.7. It was assumed that the AdminAPI
no longer supports 5.7, so error handling for this scenario was not
implemented. This oversight compromised the preconditions check, leading
to the unexpected error.

This patch ensures proper error handling in the Locking Service
checkers, preventing this issue.

Change-Id: I7dac742574a70c138a40578fd6db60d3b53f16aa