BUG#37028264 Segmentation fault executing clusterset.status command
In cases where all members of a Replica Cluster become unreachable
except for the Read-Replicas, running ClusterSet.status() would result
in a crash.

This regression was introduced by WL#16408, which added replication
compatibility checks but didn't account for this scenario. The crash
occurs when attempting to retrieve the primary member's credentials for
establishing a session to a potential source of the Read-Replica and
run the replication compatibility checks.

This patch prevents the crash by skipping the check when the required
sessions cannot be established.

Change-Id: Ic1a15a71ad455d9d1de3be75df8e5f76575b4a99