BUG#37004769 Add instance to primary cluster of a clusterset failing when all replica clusters are down
When adding a new member to the Primary Cluster of a ClusterSet, an
error occurs if one or more Replica Clusters are unreachable.
The error reports that the Replica Cluster <name> is unreachable.

This regression was introduced by WL#16408, which added replication
compatibility checks but did not account for this scenario. Specifically,
if the target Cluster is the Primary Cluster, the checks verify whether
the instance being added is compatible with all active members of all
Replica Clusters. However, unreachable Replica Clusters should be
ignored during this process, but they were not.

This patch resolves the issue by ensuring that unreachable Replica
Clusters are properly ignored during the compatibility checks.

Change-Id: Id5881b43bad12b0b86bc302dfc6c23ace46726b2