BUG#36657628 addInstance: failed clone and 'group_replication_transaction_size_limit' setting
When adding an instance to a Cluster using clone-based recovery, if the
clone process fails for any reason, the command leaves both the
instance and the Cluster in an inconsistent state. Specifically, Group
Replication remains running on the instance in an erroneous state, and
Cluster.status() reports multiple errors related to metadata and
configuration. It also suggests using Cluster.rescan() to resolve the
issues.

However, it becomes impossible to repair the Cluster's state using
Cluster.rescan() and to remove the instance from the Cluster using
.remove_instance(). The only workaround is to shut down the instance
and wait for Group Replication to expel it from the group. Once that
happens, Cluster.rescan() can be used to clean up any undesired
artifacts, such as unused recovery accounts.

The root cause of this issue was that add_instance() did not handle
errors from the clone process correctly and did not to stop Group
Replication when needed. In particular, the errors happening after Group
Replication was already started.

This patch resolves the issue by ensuring that add_instance() properly
handles clone process failures. It fully reverts the state of the
target instance, including stopping Group Replication, to prevent the
Cluster from entering an invalid state.

Change-Id: Ie9d97d000baa46c8af776b9fa8747cfd989cb770