BUG#36591845 rescan command help should mention group_replication_transaction_size_limit
The .rescan() command overwrites the existing value of
'group_replication_transaction_size_limit' if needed and it does it
silently. While there's no explicit option for it, this is not
mentioned in the help text.

Additionally, several operations performed by the command were not
documented.

This patch updates the command's help text to accurately describe and
document all operations performed by .rescan().

Change-Id: Ib5969455108edb41298f78ffd7ca0560f88462f6