BUG#36836320 Python-based credential process fails because PYTHONHOME is modified by mysqlsh
As a fix for BUG#36200851, Shell was setting the PYTHONHOME environment
variable, which allowed the bundled Python executable to find its
libraries. This variable was set regardless of the OS, and even if
Python was not bundled, even though that issue only affected Windows
packages.

With this fix, PYTHONHOME is no longer set. Instead, the bundled python
executable on Windows has been moved to another directory, creating a
directory structure that corresponds to the structure used by a
standalone Python installation.

Change-Id: Id4998d2a88f24c3766383c15cc1de024372e6d98