BUG#36614218 Cluster.addReplicaInstance() should detect duplicate in replicationSources list
When duplicate values are provided in 'replicationSources', it causes a
generic and unexpected MySQL error, leading to the failure of the
command. For example, using duplicate entries in 'replicationSources'
in .add_replica_instance() results in the following error:

.. asynchronous_connection_failover_add_source UDF failed; Error
inserting row to the table. (MySQL Error 3200)

While the MySQL error is technically correct, the AdminAPI should
proactively prevent this by disallowing duplicate values in
'replicationSources', as duplicates are invalid in this context.

This patch addresses the issue by introducing a check for duplicate
entries in 'replicationSources', ensuring the command fails early with
a clear and informative error message. The check has been added to the
following commands:

  - .add_replica_instance(x, {"replicationSources": Y})
  - .set_instance_option(x, "replicationSources", Y)
  - .rejoin_instance(x)

Change-Id: Ib8f86aa412b86b4da1faabd10984c2140762b89e