Misc Documentation Fixes BUG#35912863 BUG#36731623
BUG#35912863 Contribution: fix: String variable type

There was a typo in the return value for ClassicResult.getStatementId.

BUG#36731623 C++ code showing up in API docs for PY/JS

The problem was caused by doxigen formatted documentation in the
header file for shell.register_sql_handler as well as for missing
full specification of the function for PY/JS in the cc file.

Change-Id: I1936f23ada90433316d91d1741f0c778a811f839