BUG#36339280 interactive shell sessions should be marked interactive
This patch updates the connection logic on classic sessions in a way
that the CLIENT_CAN_HANDLE_EXPIRED_PASSWORDS and CLIENT_INTERACTIVE
options are only turned ON if the session is interactive.

Additionally introduces a new connection option "client-interactive"
so the session can be set interactive/non interactive on demand.

Change-Id: I4f81c41d56c4dc52f77f90dfef5516dd163d89ff