BUG#36054619 BUG#36682741
When adding or rejoining a member to a Cluster, InnoDB Clone can be
used to provision the instance. However, it's not the Shell that
triggers the Clone process; instead, Group Replication handles this,
and it can be forced by setting group_replication_clone_threshold=1.
Shell verifies which provisioning method can be used and allows forcing
clone.

Currently, the commands do not check if there are compatible clone
donors available in the Cluster, which is required for clone-based
recovery. If no compatible donors are available, Group Replication
automatically falls back to incremental recovery, which may fail if
only Clone can achieve the provisioning. In such cases, the commands
print an unexpected warning and give users the incorrect impression
that the instance was successfully added or rejoined to the Cluster
when it wasn’t.

Additionally, other commands that rely on the Shell to trigger Clone
for provisioning do not validate donor-recipient compatibility
beforehand, leading to situations where Clone may not be a viable
option although it's offered to the user as a choice.

This patch introduces a new check in <Cluster>.add_instance() and
<Cluster>.rejoin_instance() to ensure there is at least one compatible
clone donor before proceeding with provisioning and offering choices to
the user. The following commands have also been updated to perform
earlier the compatibility checks between donor and recipient to provide
the correct prompts and outcomes:

  - <ReplicaSet>.add_instance()
  - <ReplicaSet>.rejoin_instance()
  - <Cluster>.add_replica_instance()
  - <Cluster>.rejoin_instance() (Read-Replicas)
  - <ClusterSet>.create_replica_cluster()

For Clone to be considered compatible between a donor and recipient, the
following criteria must be met:

  - Both must be running on the same operating system.
  - Both must be on the same platform (Windows, macOS, *nix).
  - The versions must be compatible:
    + Both must be >= 8.0.17.
    + If both are >= 8.0.37, only the major and minor versions need to
      match (e.g., 8.1.4 and 8.1.10).
    + For versions >= 8.0.17 and < 8.0.37, the major, minor, and patch
      numbers must match.

This patch also fixes a bug in .add_replica_cluster() and
.rejoin_instance() (for read-replicas) when recoveryMethod is set to
incremental and mixed versions are used. Previously, the commands
erroneously checked for Clone compatibility in such scenarios, causing
errors because Clone does not support mixed-version environments. In
other words, the checks for clone compatibility shouldn't be executed in
that case.

Change-Id: I827b46ef3ceb3b37ea0930bc562c5c954af7f6d1